Move API handling to common location in codebase
- These are client-side actions only and thus need to include sensible
  error handling
- I've left the fetchers.ts file for requests made statically since I
  think it'd be confusing to mix them in together
- Went with `asyncs` to match `hooks` (since we have both `useHook` and
  `useAsync`)
