+++ frontend/pages/login/[service].tsx
@@ -5,8 +5,8 @@ import { NextSeo } from "next-seo"
 import { useRouter } from "next/router"
 import { useEffect, useState } from "react"
 import { toast } from "react-toastify"
+import { login } from "../../src/asyncs/login"
 import Spinner from "../../src/components/Spinner"
-import { login } from "../../src/context/actions"
 import { useUserContext, useUserDispatch } from "../../src/context/user-info"
 import { fetchLoginProviders } from "../../src/fetchers"
 import { usePendingTransaction } from "../../src/hooks/usePendingTransaction"
