+++ frontend/src/components/payment/DonationInput.tsx
@@ -2,49 +2,12 @@ import { useTranslation } from "next-i18next"
 import Router from "next/router"
 import React, { FormEvent, FunctionComponent, useEffect, useState } from "react"
 import { toast } from "react-toastify"
-import { TRANSACTIONS_URL } from "../../env"
+import { initiateDonation } from "../../asyncs/payment"
 import Button from "../Button"
 import Spinner from "../Spinner"
 import styles from "./DonationInput.module.scss"
 
 const minDonation = 5
-
-async function initiateDonation(recipient, amount: number): Promise<string> {
-  let res: Response
-  try {
-    res = await fetch(TRANSACTIONS_URL, {
-      method: "POST",
-      credentials: "include",
-      headers: {
-        "Content-Type": "application/json",
-      },
-      body: JSON.stringify({
-        summary: {
-          value: amount,
-          currency: "usd",
-          kind: "donation",
-        },
-        details: [
-          {
-            recipient,
-            amount,
-            currency: "usd",
-            kind: "donation",
-          },
-        ],
-      }),
-    })
-  } catch {
-    throw "network-error-try-again"
-  }
-
-  if (res.ok) {
-    const data = await res.json()
-    return data.id
-  } else {
-    throw "network-error-try-again"
-  }
-}
 
 interface Props {
   org: string
