+++ frontend/src/components/payment/checkout/Checkout.tsx
@@ -1,34 +1,12 @@
-import { useTranslation } from "next-i18next"
 import { useRouter } from "next/router"
 import { FunctionComponent, ReactElement, useEffect, useState } from "react"
-import { WALLET_INFO_URL } from "../../../env"
+import { getPaymentCards } from "../../../asyncs/payment"
 import { PaymentCard, TransactionDetailed } from "../../../types/Payment"
 import Spinner from "../../Spinner"
 import TransactionCancelButton from "../transactions/TransactionCancelButton"
 import CardSelect from "./CardSelect"
 import styles from "./Checkout.module.scss"
 import PaymentForm from "./PaymentForm"
-
-/**
- * Performs API request to retrieve details of all cards the user has saved
- * @returns array of saved cards
- */
-async function getCards() {
-  let res: Response
-  try {
-    res = await fetch(WALLET_INFO_URL, { credentials: "include" })
-  } catch {
-    throw "failed-to-load-refresh"
-  }
-
-  if (res.ok) {
-    // Not checking status, server only complains if not logged in (which we enforce)
-    const data = await res.json()
-    return data.cards
-  } else {
-    throw "failed-to-load-refresh"
-  }
-}
 
 interface Props {
   transaction: TransactionDetailed
@@ -44,7 +22,6 @@ enum Stage {
 const detailsPage = `${process.env.NEXT_PUBLIC_SITE_BASE_URI}/payment/details`
 
 const Checkout: FunctionComponent<Props> = ({ transaction, clientSecret }) => {
-  const { t } = useTranslation()
   const router = useRouter()
 
   const [currentStage, setStage] = useState(Stage.Loading)
@@ -55,7 +32,7 @@ const Checkout: FunctionComponent<Props> = ({ transaction, clientSecret }) => {
 
   // Cards should only be retrieved once
   useEffect(() => {
-    getCards()
+    getPaymentCards()
       .then((data) => {
         // User may have no saved cards to select from
         if (data.length) {
