+++ frontend/src/components/payment/transactions/TransactionCancelButton.tsx
@@ -1,28 +1,8 @@
 import { useTranslation } from "next-i18next"
 import { FunctionComponent, useEffect, useState } from "react"
 import { toast } from "react-toastify"
-import { TRANSACTION_CANCEL_URL } from "../../../env"
+import { cancelTransaction } from "../../../asyncs/payment"
 import Button from "../../Button"
-
-/**
- * Performs API request to cancel an active transaction
- * @param transactionId ID of the transaction to cancel
- */
-async function cancelTransaction(transactionId: string) {
-  let res: Response
-  try {
-    res = await fetch(TRANSACTION_CANCEL_URL(transactionId), {
-      method: "POST",
-      credentials: "include",
-    })
-  } catch {
-    throw "network-error-try-again"
-  }
-
-  if (!res.ok) {
-    throw "network-error-try-again"
-  }
-}
 
 interface Props {
   id: string
